Create README.md
Implemented a JavaScript based Grid-Box Click game. The functionality includes randomly lighting up the square boxes within a grid upon clicking on the start button. When a lit square is clicked, the game records the response time and the number of attempts it gets clicked and also it notes how much time it required to click the particular lighten grid. Furthermore, the game also had a reset and pause buttons to clear and pause both the grid and the response table.